refactor: Update Dockerfile and requirements.txt to include additional Neuron-X dependencies
This commit updates the Dockerfile and requirements.txt files in the inferentia directory. It adds the following Neuron-X dependencies:
- libneuronxla
- aws-neuronx-runtime-discovery
- neuronx-distributed

Additionally, the commit upgrades the neuronx-cc version to 2.* and installs the optimum package with the "neuron" extra.

These changes are necessary to ensure compatibility with the Neuron-X SDK and improve code organization.